fix: Change DefaultExecutionOrder attributes order (shot.MinValue -> int.MinValue)

+++ Core/Runtime/UnityExtensions/HisaCat/CoroutineAction/CoroutineAction.cs
@@ -5,7 +5,7 @@ using UnityEngine.Networking;
 
 namespace HisaCat
 {
-    [DefaultExecutionOrder(short.MinValue)]
+    [DefaultExecutionOrder(int.MinValue)]
     public class CoroutineAction : MonoBehaviour
     {
 #if UNITY_EDITOR
